+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/CDU/A320_Neo_CDU_PerformancePage.js
@@ -72,15 +72,28 @@ class CDUPerformancePage {
         let v1 = "---";
         let vR = "---";
         let v2 = "---";
+        let v1Check = "{small}\xa0\xa0\xa0\xa0{end}";
+        let vRCheck = "{small}\xa0\xa0\xa0\xa0{end}";
+        let v2Check = "{small}\xa0\xa0\xa0\xa0{end}";
         if (mcdu.currentFlightPhase < FlightPhase.FLIGHT_PHASE_TAKEOFF) {
             v1 = "{amber}___{end}";
             if (mcdu.v1Speed) {
-                v1 = `{cyan}${("" + mcdu.v1Speed).padEnd(3)}{end}`;
+                if (mcdu._v1Checked) {
+                    v1 = `{cyan}${("" + mcdu.v1Speed).padEnd(3)}{end}`;
+                } else {
+                    v1Check = `{small}{cyan}\xa0${("" + mcdu.v1Speed).padEnd(3)}{end}{end}`;
+                }
             }
             mcdu.onLeftInput[0] = (value) => {
                 if (value === "") {
-                    // not real: v-speed helper
-                    mcdu.sendDataToScratchpad(mcdu._getV1Speed().toString());
+                    if (mcdu._v1Checked) {
+                        // not real: v-speed helper
+                        mcdu.sendDataToScratchpad(mcdu._getV1Speed().toString());
+                    } else {
+                        mcdu._v1Checked = true;
+                        mcdu.tryRemoveMessage(NXSystemMessages.checkToData.text);
+                        mcdu.vSpeedDisagreeCheck();
+                    }
                     CDUPerformancePage.ShowTAKEOFFPage(mcdu);
                 } else {
                     if (mcdu.trySetV1Speed(value)) {
@@ -90,11 +103,21 @@ class CDUPerformancePage {
             };
             vR = "{amber}___{end}";
             if (mcdu.vRSpeed) {
-                vR = `{cyan}${("" + mcdu.vRSpeed).padEnd(3)}{end}`;
+                if (mcdu._vRChecked) {
+                    vR = `{cyan}${("" + mcdu.vRSpeed).padEnd(3)}{end}`;
+                } else {
+                    vRCheck = `{small}{cyan}\xa0${("" + mcdu.vRSpeed).padEnd(3)}{end}{end}`;
+                }
             }
             mcdu.onLeftInput[1] = (value) => {
                 if (value === "") {
-                    mcdu.sendDataToScratchpad(mcdu._getVRSpeed().toString());
+                    if (mcdu._vRChecked) {
+                        mcdu.sendDataToScratchpad(mcdu._getVRSpeed().toString());
+                    } else {
+                        mcdu._vRChecked = true;
+                        mcdu.tryRemoveMessage(NXSystemMessages.checkToData.text);
+                        mcdu.vSpeedDisagreeCheck();
+                    }
                     CDUPerformancePage.ShowTAKEOFFPage(mcdu);
                 } else {
                     if (mcdu.trySetVRSpeed(value)) {
@@ -104,11 +127,21 @@ class CDUPerformancePage {
             };
             v2 = "{amber}___{end}";
             if (mcdu.v2Speed) {
-                v2 = `{cyan}${("" + mcdu.v2Speed).padEnd(3)}{end}`;
+                if (mcdu._v2Checked) {
+                    v2 = `{cyan}${("" + mcdu.v2Speed).padEnd(3)}{end}`;
+                } else {
+                    v2Check = `{small}{cyan}\xa0${("" + mcdu.v2Speed).padEnd(3)}{end}{end}`;
+                }
             }
             mcdu.onLeftInput[2] = (value) => {
                 if (value === "") {
-                    mcdu.sendDataToScratchpad(mcdu._getV2Speed().toString());
+                    if (mcdu._v2Checked) {
+                        mcdu.sendDataToScratchpad(mcdu._getV2Speed().toString());
+                    } else {
+                        mcdu._v2Checked = true;
+                        mcdu.tryRemoveMessage(NXSystemMessages.checkToData.text);
+                        mcdu.vSpeedDisagreeCheck();
+                    }
                     CDUPerformancePage.ShowTAKEOFFPage(mcdu);
                 } else {
                     if (mcdu.trySetV2Speed(value)) {
@@ -255,6 +288,7 @@ class CDUPerformancePage {
             flapsThs = `${flaps}/${ths}[color]cyan`;
             mcdu.onRightInput[2] = (value) => {
                 if (mcdu.trySetFlapsTHS(value)) {
+                    mcdu.onToDataChanged();
                     CDUPerformancePage.ShowTAKEOFFPage(mcdu);
                 }
             };
@@ -272,6 +306,7 @@ class CDUPerformancePage {
             }
             mcdu.onRightInput[3] = (value) => {
                 if (mcdu.setPerfTOFlexTemp(value)) {
+                    mcdu.onToDataChanged();
                     CDUPerformancePage.ShowTAKEOFFPage(mcdu);
                 }
             };
@@ -313,28 +348,42 @@ class CDUPerformancePage {
             engOutAcc = "";
         }
 
+        let next = "NEXT\xa0";
+        let nextPhase = "PHASE>";
+        if ((!mcdu._v1Checked || !mcdu._vRChecked || !mcdu._v2Checked) && mcdu.currentFlightPhase < FlightPhase.FLIGHT_PHASE_TAKEOFF) {
+            next = "CONFIRM\xa0";
+            nextPhase = "TO DATA*";
+            mcdu.onRightInput[5] = (value) => {
+                mcdu._v1Checked = true;
+                mcdu._vRChecked = true;
+                mcdu._v2Checked = true;
+                mcdu.vSpeedDisagreeCheck();
+                CDUPerformancePage.ShowTAKEOFFPage(mcdu);
+            };
+        } else {
+            mcdu.rightInputDelay[5] = () => {
+                return mcdu.getDelaySwitchPage();
+            };
+            mcdu.onRightInput[5] = (value) => {
+                CDUPerformancePage.ShowCLBPage(mcdu);
+            };
+        }
+
         mcdu.setTemplate([
             ["TAKE OFF RWY " + runway.padStart(3, "\xa0") + "[color]" + titleColor],
             ["\xa0V1\xa0\xa0\xa0\xa0FLP RETR", ""],
-            [v1 + "\xa0\xa0\xa0\xa0\xa0F=" + flpRetrCell, ""],
+            [v1 + v1Check + "\xa0F=" + flpRetrCell, ""],
             ["\xa0VR\xa0\xa0\xa0\xa0SLT RETR", "TO SHIFT\xa0"],
-            [vR + "\xa0\xa0\xa0\xa0\xa0S=" + sltRetrCell, toShiftCell],
+            [vR + vRCheck + "\xa0S=" + sltRetrCell, toShiftCell],
             ["\xa0V2\xa0\xa0\xa0\xa0\xa0\xa0\xa0CLEAN", "FLAPS/THS"],
-            [v2 + "\xa0\xa0\xa0\xa0\xa0O=" + cleanCell, flapsThs],
+            [v2 + v2Check + "\xa0O=" + cleanCell, flapsThs],
             ["TRANS ALT", "FLEX TO TEMP"],
             [transAltCell, flexTakeOffTempCell],
             ["THR RED/ACC", "ENG OUT ACC"],
             [thrRedAcc, engOutAcc],
-            ["\xa0UPLINK[color]inop", "NEXT\xa0"],
-            ["<TO DATA[color]inop", "PHASE>"]
+            ["\xa0UPLINK[color]inop", next],
+            ["<TO DATA[color]inop", nextPhase]
         ]);
-
-        mcdu.rightInputDelay[5] = () => {
-            return mcdu.getDelaySwitchPage();
-        };
-        mcdu.onRightInput[5] = (value) => {
-            CDUPerformancePage.ShowCLBPage(mcdu);
-        };
     }
     static ShowCLBPage(mcdu, confirmAppr = false) {
         mcdu.clearDisplay();